+++ NEWS.md
@@ -1,9 +1,16 @@
 # ospsuite (development version)
 
+## Minor improvements and bug fixes
+
+- Improve printing of parameters with RHS formulas (\#1511)
+
+# ospsuite 12.2.0
+
 ## Major Changes
 
 - New `runSimulationsFromSnapshot()` to run simulations from `.json` snapshots files,
 - New `convertSnapshot()` to convert project snapshots between `.json` and `.pksim5` files.
+- New `getMolWeightFor()` to retrieve molecular weight for molecule of `Quantity`.
 
 # ospsuite 12.1.0